added work on faqs

+++ app/Http/Livewire/Admin/Users/Delete.php
@@ -26,7 +26,9 @@ class Delete extends Component
     public function delete()
     {
         try {
-            unlink('public/' . $this->user->profile);
+            if(isset($this->user->profile)){
+                unlink('public/' . $this->user->profile);
+            }
         } catch (\Exception $e) {
             $this->emit('error', "An error occurred while deleting");
             return;
